move validations from services to validations
